Rename Permutations of String to permutationsofstringByRitikaJain
Given string str, the task is to print all the permutations of str. A permutation is an arrangement of all or part of a set of objects, with regard to the order of the arrangement. For instance, the words ‘bat’ and ‘tab’ represent two distinct permutation (or arrangements) of a similar three-letter word.

I have solved this problem using Backtracking.